refactor: migrate the types to types file
